requirements.txt
certifi==2022.5.18.1
charset-normalizer==2.0.12
idna==3.3
pathlib==1.0.1
ply==3.11
psycopg2>=2.7
pycryptodome==3.14.1
pyTelegramBotAPI==4.5.1
requests==2.27.1
six==1.16.0
stone==3.3.1
tenacity==5.1.5
urllib3==1.26.9
certifi==2022.5.18.1
charset-normalizer==2.0.12
idna==3.3
pathlib==1.0.1
ply==3.11
protobuf==4.21.12
psycopg2==2.9.5
pycryptodome==3.14.1
pyTelegramBotAPI==4.5.1
requests==2.27.1
six==1.16.0
stone==3.3.1
tenacity==5.1.5
urllib3==1.26.9
certifi==2022.5.18.1
charset-normalizer==2.0.12
idna==3.3
mysql-connector-python==8.0.32
pathlib==1.0.1
ply==3.11
protobuf==3.20.3
psycopg2==2.9.5
pycryptodome==3.14.1
pyTelegramBotAPI==4.5.1
requests==2.27.1
six==1.16.0
stone==3.3.1
tenacity==5.1.5
urllib3==1.26.9
aiohttp==3.8.3
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.2.0
certifi==2022.5.18.1
charset-normalizer==2.0.12
colorama==0.4.6
frozenlist==1.3.3
idna==3.3
multidict==6.0.4
mysql-connector-python==8.0.32
openai==0.26.4
pathlib==1.0.1
ply==3.11
protobuf==3.20.3
psycopg2==2.9.5
pycryptodome==3.14.1
pyTelegramBotAPI==4.5.1
requests==2.27.1
six==1.16.0
stone==3.3.1
tenacity==5.1.5
tqdm==4.64.1
urllib3==1.26.9
yarl==1.8.2
